# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API task_master_api",
        "description": "Documentación de la API para la gestión de tareas.",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://task_master_api.test/api",
            "description": "Servidor API principal"
        },
        {
            "url": "http://task_master_api.test/api",
            "description": "Servidor API principal"
        }
    ],
    "paths": {
...
        "/api/v1/tasks/me": {
            "get": {
                "tags": [
                    "Tareas"
                ],
                "summary": "Obtener lista de tareas del usuario autenticado",
                "description": "Retorna una lista de tareas asociadas al usuario autenticado, requiere un token JWT en los headers.",
                "operationId": "c02a2365a23f8778d445470dc18ba37d",
                "parameters": [
                    {
                        "name": "token",
                        "in": "header",
                        "description": "Token JWT del usuario autenticado",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de tareas obtenida correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Lista de tareas obtenidas correctamente"
                                        },
                                        "tasks": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Task"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuario no autorizado o no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Usuario no encontrado"
                                        },
                                        "message_detail": {
                                            "type": "string",
                                            "example": "No fue posible encontrar al usuario logueado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error al listar tareas"
                                        },
                                        "message_detail": {
                                            "type": "string",
                                            "example": "Detalles del error interno"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Task": {
                "title": "Tarea",
                "description": "Estructura de una tarea",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Deserunt ut."
                    },
                    "description": {
                        "type": "string",
                        "example": "Descripción de la tarea."
                    },
                    "expiration_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-03-29"
                    },
                    "status": {
                        "type": "string",
                        "example": "Pendiente"
                    },
                    "user": {
                        "type": "string",
                        "example": "Lucy West"
                    }
                },
                "type": "object"
            }
        }
    },
...
}